pyproject.toml
[project]
name = "mainproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.135.3",
    "loguru>=0.7.3",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.43.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.9",
]

# uv run ruff check --fix
# uv run ruff format --check
# uv run ruff format src tests
[tool.ruff]
# Где лежит исходный код
src = ["src", "tests"]

# Папки, которые Ruff полностью игнорирует
exclude = [
    "alembic/versions/*",   # ← важно! миграции Alembic не трогаем
    ".venv",
    ".git",
    "__pycache__",
    "logs",
    "*.log",
    "alembic.ini",
    "Dockerfile",
    "docker-compose.yml",
]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C4", "UP"]
ignore = [
    "E501",      # длинные строки
    "UP017",     # ← добавь эту строку (datetime.UTC)
]

[tool.ruff.format]
quote-style = "double"
